Explanation
The simulation software used is MatLab. The dynamic simulations depicted are computed by the ordinary differential equation solver. In this sense, this section is devoted to describing the process of simulation with the aid of ODES of MatLab.

The mathematical models described are programmed in the MatLab editor with the aid of ODES and declaring functions the system is simulated. Then, the linearization of the entire system is done using \textit{fsolve} of MatLab. 

The complete system shown in Figure 1 is integrated into a function where all the dynamic models are declared as a function of themselves. The following .txt files capture the code of the complete system programmed, simulated, and solved in the editor of MatLab.

The ordinary differential equation solver used to simulate the entire system is the {ode23tb} as shown in line 15 of the main code, in this line, the function {System} contains all the models programmed as function; the input and output are indicated. In this file, the main code encompasses all the programmed differential equations, where the variable {t} records the vector of time of the entire simulation and the variable {y} records the entire vector of the state variables.